+++ NEWS.md
@@ -1,3 +1,8 @@
+### packageRank 0.9.7.9003
+
+- add rPlot(axis.package, axis.package.version) and 
+  rTotPlot(axis.package, axis.package.version).
+
 ### packageRank 0.9.7.9002
 
 - add cranPlot(axis.package, axis.package.version).